% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{ceride_palettes}
\alias{ceride_palettes}
\title{A collection of palettes drawn from data visualizations in Cerîde-i Adliyye (The Justice Gazette), published in Turkey in the 1920s.
Thanks to Jilly MacKay}
\format{
An object of class \code{list} of length 6.
}
\usage{
ceride_palettes
}
\description{
A collection of palettes drawn from data visualizations in Cerîde-i Adliyye (The Justice Gazette), published in Turkey in the 1920s.
Thanks to Jilly MacKay
}
\keyword{datasets}
