% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_palette.R
\name{ceride_show_palette}
\alias{ceride_show_palette}
\title{Show a given colour palette from Ceride-i Adliyye scheme}
\usage{
ceride_show_palette(palette)
}
\arguments{
\item{palette}{vector of character hex RGB values}
}
\description{
Given a character vector (hex RGB values), display palette in graphics window.
}
\examples{
library(ceride.i.adliyye)
nord_show_palette("four_colors")
}
