% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceride_scale.R
\name{make_ceride_palette}
\alias{make_ceride_palette}
\title{Creates a palette from your specified colours}
\usage{
make_ceride_palette(palette = "four_colors", alpha = 1, reverse = FALSE)
}
\arguments{
\item{palette}{Choose from 'ceride_palettes' list}

\item{alpha}{transparency}

\item{reverse}{If TRUE, the direction of the colours is reversed.}
}
\description{
Creates a palette from your specified colours
}
