% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceride_scale.R
\name{scale_color_ceride}
\alias{scale_color_ceride}
\title{Creates a colour scale for ggplot2}
\usage{
scale_color_ceride(
  palette = "five_colors",
  discrete = TRUE,
  alpha = 1,
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{Choose from 'ceride_palettes' list}

\item{discrete}{whether to use a discrete colour palette}

\item{alpha}{transparency}

\item{reverse}{logical, Reverse the order of the colours?}

\item{...}{additional arguments to pass to scale_color_gradientn}
}
\description{
Creates a colour scale for ggplot2
}
\examples{
library(ggplot2)
library(ceride.i.adliyye)

ggplot(mpg) +
 geom_point(aes(x = hwy, y = cty, color = manufacturer)) +
 scale_color_ceride("five_colors")

}
