% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceride_scale.R
\name{scale_fill_ceride}
\alias{scale_fill_ceride}
\title{Creates a  fill scale for ggplot2}
\usage{
scale_fill_ceride(
  palette = "five_colors",
  discrete = TRUE,
  alpha = 1,
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{Choose from 'ceride_palettes' list}

\item{discrete}{whether to use a discrete colour palette}

\item{alpha}{The alpha transparency, a number in [0,1], see argument alpha in
\code{\link[grDevices]{hsv}}.}

\item{...}{additional arguments to pass to scale_color_gradientn}
}
\description{
Creates a  fill scale for ggplot2
}
\examples{
library(ggplot2)
library(ceride.i.adliyye)

ggplot(aes(x = manufacturer, fill = manufacturer), data = mpg) +
  geom_bar() +
  scale_fill_ceride()
}
